feat(loading-button): make the async icon button more generic
